+++ src/ext/java/org/opentripplanner/ext/siri/updater/SiriSXUpdater.java
@@ -1,10 +1,13 @@
 package org.opentripplanner.ext.siri.updater;
 
+import java.time.Duration;
 import java.time.ZonedDateTime;
 import java.util.UUID;
 import org.opentripplanner.ext.siri.SiriAlertsUpdateHandler;
 import org.opentripplanner.ext.siri.SiriFuzzyTripMatcher;
 import org.opentripplanner.framework.io.OtpHttpClientException;
+import org.opentripplanner.framework.retry.OtpRetry;
+import org.opentripplanner.framework.retry.OtpRetryBuilder;
 import org.opentripplanner.routing.impl.TransitAlertServiceImpl;
 import org.opentripplanner.routing.services.TransitAlertService;
 import org.opentripplanner.transit.service.DefaultTransitService;
@@ -20,7 +23,10 @@ import uk.org.siri.siri20.Siri;
 public class SiriSXUpdater extends PollingGraphUpdater implements TransitAlertProvider {
 
   private static final Logger LOG = LoggerFactory.getLogger(SiriSXUpdater.class);
-  private static final long RETRY_INTERVAL_MILLIS = 5000;
+  private static final int RETRY_MAX_ATTEMPTS = 3;
+  private static final Duration RETRY_INITIAL_DELAY = Duration.ofSeconds(5);
+  private static final int RETRY_BACKOFF = 2;
+
   private final String url;
   private final String originalRequestorRef;
   private final TransitAlertService transitAlertService;
@@ -28,8 +34,12 @@ public class SiriSXUpdater extends PollingGraphUpdater implements TransitAlertPr
   private WriteToGraphCallback saveResultOnGraph;
   private ZonedDateTime lastTimestamp = ZonedDateTime.now().minusWeeks(1);
   private String requestorRef;
+  /**
+   * Global retry counter used to create a new unique requestorRef after each retry.
+   */
   private int retryCount = 0;
   private final SiriHttpLoader siriHttpLoader;
+  private final OtpRetry retry;
 
   public SiriSXUpdater(SiriSXUpdaterParameters config, TransitModel transitModel) {
     super(config);
@@ -55,6 +65,16 @@ public class SiriSXUpdater extends PollingGraphUpdater implements TransitAlertPr
       );
     siriHttpLoader = new SiriHttpLoader(url, config.timeout(), config.requestHeaders());
 
+    retry =
+      new OtpRetryBuilder()
+        .withName("SIRI-SX Update")
+        .withMaxAttempts(RETRY_MAX_ATTEMPTS)
+        .withInitialRetryInterval(RETRY_INITIAL_DELAY)
+        .withBackoffMultiplier(RETRY_BACKOFF)
+        .withRetryableException(OtpHttpClientException.class::isInstance)
+        .withOnRetry(this::updateRequestorRef)
+        .build();
+
     LOG.info(
       "Creating real-time alert updater (SIRI SX) running every {} seconds : {}",
       pollingPeriod(),
@@ -76,38 +96,30 @@ public class SiriSXUpdater extends PollingGraphUpdater implements TransitAlertPr
   }
 
   @Override
-  protected void runPolling() throws InterruptedException {
-    try {
-      boolean moreData = false;
-      do {
-        Siri updates = getUpdates();
-        if (updates != null) {
-          ServiceDelivery serviceDelivery = updates.getServiceDelivery();
-          moreData = Boolean.TRUE.equals(serviceDelivery.isMoreData());
-          // Mark this updater as primed after last page of updates. Copy moreData into a final
-          // primitive, because the object moreData persists across iterations.
-          final boolean markPrimed = !moreData;
-          if (serviceDelivery.getSituationExchangeDeliveries() != null) {
-            saveResultOnGraph.execute((graph, transitModel) -> {
-              updateHandler.update(serviceDelivery);
-              if (markPrimed) primed = true;
-            });
-          }
+  protected void runPolling() {
+    retry.execute(this::updateSiri);
+  }
+
+  private void updateSiri() {
+    boolean moreData = false;
+    do {
+      Siri updates = getUpdates();
+      if (updates != null) {
+        ServiceDelivery serviceDelivery = updates.getServiceDelivery();
+        moreData = Boolean.TRUE.equals(serviceDelivery.isMoreData());
+        // Mark this updater as primed after last page of updates. Copy moreData into a final
+        // primitive, because the object moreData persists across iterations.
+        final boolean markPrimed = !moreData;
+        if (serviceDelivery.getSituationExchangeDeliveries() != null) {
+          saveResultOnGraph.execute((graph, transitModel) -> {
+            updateHandler.update(serviceDelivery);
+            if (markPrimed) {
+              primed = true;
+            }
+          });
         }
-      } while (moreData);
-    } catch (OtpHttpClientException e) {
-      final long sleepTime = RETRY_INTERVAL_MILLIS + RETRY_INTERVAL_MILLIS * retryCount;
-
-      retryCount++;
-
-      LOG.info("Caught timeout - retry no. {} after {} millis", retryCount, sleepTime);
-
-      Thread.sleep(sleepTime);
-
-      // Creating new requestorRef so all data is refreshed
-      requestorRef = originalRequestorRef + "-retry-" + retryCount;
-      runPolling();
-    }
+      }
+    } while (moreData);
   }
 
   private Siri getUpdates() {
@@ -129,13 +141,30 @@ public class SiriSXUpdater extends PollingGraphUpdater implements TransitAlertPr
       lastTimestamp = responseTimestamp;
       return siri;
     } catch (OtpHttpClientException e) {
-      LOG.info("Failed after {} ms", (System.currentTimeMillis() - t1));
-      LOG.error("Error reading SIRI feed from " + url, e);
+      LOG.info(
+        "Retryable exception while reading SIRI feed from {} after {} ms",
+        url,
+        (System.currentTimeMillis() - t1)
+      );
       throw e;
     } catch (Exception e) {
-      LOG.info("Failed after {} ms", (System.currentTimeMillis() - t1));
-      LOG.error("Error reading SIRI feed from " + url, e);
+      LOG.error(
+        "Non-retryable exception while reading SIRI feed from {} after {} ms",
+        url,
+        (System.currentTimeMillis() - t1)
+      );
     }
     return null;
+  }
+
+  /**
+   * Reset the session with the SIRI-SX server by creating a new unique requestorRef. This is
+   * required if a network error causes a request to fail and let the session in an undetermined
+   * state. Using a new requestorRef will force the SIRI-SX server to send again all available
+   * messages.
+   */
+  private void updateRequestorRef() {
+    retryCount++;
+    requestorRef = originalRequestorRef + "-retry-" + retryCount;
   }
 }