+++ src/ext/java/org/opentripplanner/ext/transmodelapi/model/plan/ItineraryFiltersInputType.java
@@ -11,8 +11,7 @@ import org.opentripplanner.ext.transmodelapi.model.EnumTypes;
 import org.opentripplanner.ext.transmodelapi.support.DataFetcherDecorator;
 import org.opentripplanner.ext.transmodelapi.support.GqlUtil;
 import org.opentripplanner.routing.algorithm.filterchain.api.TransitGeneralizedCostFilterParams;
-import org.opentripplanner.routing.api.request.framework.DoubleAlgorithmFunction;
-import org.opentripplanner.routing.api.request.framework.RequestFunctions;
+import org.opentripplanner.routing.api.request.framework.CostLinearFunction;
 import org.opentripplanner.routing.api.request.preference.ItineraryFilterPreferences;
 
 public class ItineraryFiltersInputType {
@@ -84,7 +83,7 @@ public class ItineraryFiltersInputType {
             "is set to: 3 600 + 2 * 10 000 = 26 600 plus half of the time between either departure" +
             " or arrival times of the itinerary. " +
             "Default: {\"costLimitFunction\": " +
-            RequestFunctions.serialize(dft.transitGeneralizedCostLimit().costLimitFunction()) +
+            dft.transitGeneralizedCostLimit().costLimitFunction().serialize() +
             ", \"intervalRelaxFactor\": " +
             dft.transitGeneralizedCostLimit().intervalRelaxFactor() +
             "}"
@@ -182,7 +181,7 @@ public class ItineraryFiltersInputType {
       (Map<String, ?> v) ->
         builder.withTransitGeneralizedCostLimit(
           new TransitGeneralizedCostFilterParams(
-            (DoubleAlgorithmFunction) v.get("costLimitFunction"),
+            (CostLinearFunction) v.get("costLimitFunction"),
             (double) v.get("intervalRelaxFactor")
           )
         )