+++ src/main/java/org/opentripplanner/api/mapping/PlannerErrorMapper.java
@@ -34,6 +34,9 @@ public class PlannerErrorMapper {
         } else if (domain.inputField.equals(InputField.TO_PLACE)) {
           api = new PlannerError(Message.GEOCODE_TO_NOT_FOUND);
           api.setMissing(List.of(domain.inputField.name()));
+        } else if (domain.inputField.equals(InputField.INTERMEDIATE_PLACE)) {
+          api = new PlannerError(Message.GEOCODE_INTERMEDIATE_NOT_FOUND);
+          api.setMissing(List.of(domain.inputField.name()));
         } else {
           throw new IllegalArgumentException();
         }