+++ src/main/java/org/opentripplanner/updater/vehicle_rental/datasources/VehicleRentalDataSourceFactory.java
@@ -2,20 +2,30 @@ package org.opentripplanner.updater.vehicle_rental.datasources;
 
 import org.opentripplanner.ext.smoovebikerental.SmooveBikeRentalDataSource;
 import org.opentripplanner.ext.smoovebikerental.SmooveBikeRentalDataSourceParameters;
+import org.opentripplanner.framework.io.OtpHttpClient;
 import org.opentripplanner.updater.vehicle_rental.datasources.params.GbfsVehicleRentalDataSourceParameters;
 import org.opentripplanner.updater.vehicle_rental.datasources.params.VehicleRentalDataSourceParameters;
 
 public class VehicleRentalDataSourceFactory {
 
-  public static VehicleRentalDatasource create(VehicleRentalDataSourceParameters source) {
+  public static VehicleRentalDatasource create(
+    VehicleRentalDataSourceParameters source,
+    OtpHttpClient otpHttpClient
+  ) {
     return switch (source.sourceType()) {
       // There used to be a lot more updaters and corresponding config variables here, but since
       // the industry has largely moved towards GBFS, they were removed.
       // See: https://groups.google.com/g/opentripplanner-users/c/0P8UGBJG-zE/m/xaOZnL3OAgAJ
       // If you want to add your provider-specific updater, you can move them into a sandbox module
       // and become the point of contact for the community.
-      case GBFS -> new GbfsVehicleRentalDataSource((GbfsVehicleRentalDataSourceParameters) source);
-      case SMOOVE -> new SmooveBikeRentalDataSource((SmooveBikeRentalDataSourceParameters) source);
+      case GBFS -> new GbfsVehicleRentalDataSource(
+        (GbfsVehicleRentalDataSourceParameters) source,
+        otpHttpClient
+      );
+      case SMOOVE -> new SmooveBikeRentalDataSource(
+        (SmooveBikeRentalDataSourceParameters) source,
+        otpHttpClient
+      );
     };
   }
 }