+++ src/test/java/org/opentripplanner/raptor/moduletests/D03_UnpreferredRouteTest.java
@@ -5,7 +5,6 @@ import static org.opentripplanner.raptor._data.api.PathUtils.pathsToString;
 import static org.opentripplanner.raptor._data.transit.TestRoute.route;
 import static org.opentripplanner.raptor._data.transit.TestTripPattern.pattern;
 import static org.opentripplanner.raptor._data.transit.TestTripSchedule.schedule;
-import static org.opentripplanner.routing.algorithm.raptoradapter.transit.cost.PatternCostCalculator.DEFAULT_ROUTE_RELUCTANCE;
 
 import java.util.BitSet;
 import org.junit.jupiter.api.BeforeEach;
@@ -18,8 +17,7 @@ import org.opentripplanner.raptor._data.transit.TestTripSchedule;
 import org.opentripplanner.raptor.api.request.RaptorProfile;
 import org.opentripplanner.raptor.api.request.RaptorRequestBuilder;
 import org.opentripplanner.raptor.configure.RaptorConfig;
-import org.opentripplanner.routing.api.request.framework.DoubleAlgorithmFunction;
-import org.opentripplanner.routing.api.request.framework.RequestFunctions;
+import org.opentripplanner.routing.api.request.framework.CostLinearFunction;
 import org.opentripplanner.transit.model._data.TransitModelForTest;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 
@@ -35,10 +33,7 @@ public class D03_UnpreferredRouteTest implements RaptorTestConstants {
     "Walk 30s ~ A ~ BUS %s 0:01 0:02:40 ~ B ~ Walk 20s " + "[0:00:30 0:03 2m30s 0tx $%d]";
   private static final FeedScopedId ROUTE_ID_1 = TransitModelForTest.id("1");
   private static final FeedScopedId ROUTE_ID_2 = TransitModelForTest.id("2");
-  private static final DoubleAlgorithmFunction UNPREFER_COST = RequestFunctions.createLinearFunction(
-    30000,
-    DEFAULT_ROUTE_RELUCTANCE
-  );
+  private static final CostLinearFunction UNPREFERRED_COST = CostLinearFunction.of("5m + 1t");
   private final TestTransitData data = new TestTransitData();
   private final RaptorRequestBuilder<TestTripSchedule> requestBuilder = new RaptorRequestBuilder<>();
   private final RaptorService<TestTripSchedule> raptorService = new RaptorService<>(
@@ -99,7 +94,7 @@ public class D03_UnpreferredRouteTest implements RaptorTestConstants {
       }
     }
     data.mcCostParamsBuilder().unpreferredPatterns(patterns);
-    data.mcCostParamsBuilder().unpreferredCost(UNPREFER_COST);
+    data.mcCostParamsBuilder().unpreferredCost(UNPREFERRED_COST);
   }
 
   private static String expected(String route, int cost) {