+++ src/test/java/org/opentripplanner/routing/algorithm/filterchain/deletionflagger/TransitGeneralizedCostFilterTest.java
@@ -4,12 +4,13 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.opentripplanner.model.plan.Itinerary.toStr;
 import static org.opentripplanner.model.plan.TestItineraryBuilder.newItinerary;
 
+import java.time.Duration;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.framework.model.Cost;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.model.plan.PlanTestConstants;
-import org.opentripplanner.routing.algorithm.filterchain.api.TransitGeneralizedCostFilterParams;
-import org.opentripplanner.routing.api.request.framework.RequestFunctions;
+import org.opentripplanner.routing.api.request.framework.CostLinearFunction;
 
 public class TransitGeneralizedCostFilterTest implements PlanTestConstants {
 
@@ -18,7 +19,8 @@ public class TransitGeneralizedCostFilterTest implements PlanTestConstants {
     // Create a filter with f(x) = 600 + 2x, without any penalty for waiting at the beginning or end.
     // Remove itineraries with a cost equivalent of 10 minutes and twice the min itinerary cost.
     final TransitGeneralizedCostFilter subject = new TransitGeneralizedCostFilter(
-      new TransitGeneralizedCostFilterParams(RequestFunctions.createLinearFunction(600, 2.0), 0.0)
+      CostLinearFunction.of(Duration.ofMinutes(10), 2.0),
+      0.0
     );
 
     // Walk all the way, not touched by the filter even if cost(7200) is higher than transit limit.
@@ -48,7 +50,8 @@ public class TransitGeneralizedCostFilterTest implements PlanTestConstants {
     // Remove itineraries with a cost equivalent of twice the itinerary cost plus half of the
     // waiting time.
     final TransitGeneralizedCostFilter subject = new TransitGeneralizedCostFilter(
-      new TransitGeneralizedCostFilterParams(RequestFunctions.createLinearFunction(0, 2.0), 0.5)
+      CostLinearFunction.of(Cost.ZERO, 2.0),
+      0.5
     );
 
     // Walk all the way, not touched by the filter even if cost(7200) is higher than transit limit.
@@ -78,7 +81,8 @@ public class TransitGeneralizedCostFilterTest implements PlanTestConstants {
     // Remove itineraries with a cost equivalent of twice the itinerary cost plus half of the
     // waiting time.
     final TransitGeneralizedCostFilter subject = new TransitGeneralizedCostFilter(
-      new TransitGeneralizedCostFilterParams(RequestFunctions.createLinearFunction(0, 2.0), 0.5)
+      CostLinearFunction.of(Cost.ZERO, 2.0),
+      0.5
     );
 
     // Walk all the way, not touched by the filter even if cost(7200) is higher than transit limit.